Add letter placement check
Add a function that checks whether a letter
placement is legal without actually placing the letter there.
This function is called by the placement function
and can also be used from a client-facing interface,
or from other checks.

Eventually instead of calling this function everytime
the user hovers a tile with a litter, however,
a map with potential legal placements
will be sent to the client after every game move.
This keeps the number of new client-server connections
at a reasonable size.

PS: Go supports variable definition
in if statements. So hot.