Implement game creation logic
Starting implementation with the game creation logic.
Every new game is created with a new UUID
(generated by calling a system binary).
The UUID will be sent to the client and is thenafter
used to reference to this game.
Every API regarding the game needs to have the UUID included.

At Game creation, the players are added to the Game
struct and and the struct is added to a list
of games. This list contains all the currently running games.

The number of players allowed in a game is defined with
2-4.