Add library modules for config and dict api, finish turn logic
Implement a library for sending API requests
to a dictionary service for checking word validity.

Problem with dictionary API request:
intermittend seg fault.

Also add a config file and a library to read it.

Start Implementation for the logic that
applies when a player has finished placing his
letters and wants wo finish the turn and
receive points for this word.