Implement finish turn logic
Implement functionality ta validate words
and add up the points of a word
with taking account of the tile effects.

Tile arrays of words that have already been
validated and rated, are added to a
confirmedWords array so that
they are not checked multiple times.

Possible issues yet to solve:
tiles need to be locked
if, and only if, the words on them
have been confirmed.
This has to happen through a passed
reference to the tile.