Implement parts of client lib, web api fixes, tile effect
Implement parts of the GUI and the http requests on the client side.
The board with parts of the effects can already be drawn.

Requests from the client need to be stringified JSON objects
that can be decoded by the server. NOT PURE application/json
as this will result in parsing or CORS problems.

Add checks for double word effect and center tile, etc...

Todo:
I will add the legalPlacement status of a Tile
to the board map that is requested by the client
to reduce the amount of objects that have to be transported
via the network.