Major improvement in backend and frontend design
Improvements of Data structures and function design on backend
as well as communication between server and client.
It is now no more necessary to call
a legalPlacements map, as that information was added to the
tile struct and is now part of the board struct.
Also there is no more need to ever send the playerName
from client to server (except for game start),
the backend decides which player info to return,
based on which used has the current turn.

Also improvements on the frontend. Mostly on design.

Various bugfixes.