Simplify and extend client
Turns out some of the stuff I implemented is already an
implementation feature. The tile object on the client
can be dropped.

Calling the place letter api endpoint works fine,
whereas some checks on the server side are left
to be done (has a user an actual tile).

Also, letters need to be removed from their place,
before they are picked up by the user both visually and on the server,
(call removeLetter and remove letter from board or hand of player)
otherwise, it is confusing and also the legalPlacements are wrong.