Quick scripting to make game playable
Add functionalities to make the game partly playable.
Most of those functionalities are a bit of a hack
and a bit dirty.

Especially the main.js file needs major refactoring.
Most functionalitites need to be outsourced to other modules.
Especially the draw__ functions are heavily overloaded
and carry wrong responsibilities.

The jquery-ui library solves a lot of gameplay visualization issues
(like drag and drop as well as sorting tiles on the player hand).

Add a file that provides userinteraction functionalities
for non-gameplay (user interaction that does not directly interact with game elements)
(i.e. dialogs).

Errors returned from the server need to be caught and communicated to the user
or, if possible, handled by the client implementation.

There is at the moment a run time error occuring on the server side
when a word is being checked. (Index out of bounds) Needs fix.