Use local dictionary, debugging
There is no need to call a webbased dictionary service API,
instead we can use the local dictionary for checking the words,
there is also a go library for working with this dict.
The libspellcheck uses the dictionary at usr/share/dict/words
at the moment which restricts gole server to be run on unix.

A lot of debugging work. Getting the connected horizontal and vertical tiles
is buggy.