Add checks for word fetchers, fix reference err
Before retrieving the word around a tile, check if there are
actually occupied tiles adjacent in the according directions.
Add code contracts to the two word retrieval algorithms.

Fix a reference error in the GetGameByUUID Function.
The array with games was copied and its (the copy's game stuct)
reference was passed instead of the original.