Fix point commit issue
The word multiplicator that was applied to the
calculated points every time the points for a word
were processed was set to 0 by default which eradicated the
moints it was multiplated with.

Made default 1 to get rid of this bug.

TODO:
Before the first alpha version can be relesed those
things are to be completed:
- Allow letter to be moved back from bord to hand if word not yet confirmed
- Make sure player has letter in hand before played word can be confirmed
- Add new letters to player hand after played word
- Implement usage of wildcard tile