Implement remove from board functionality
Implement the server side api endpoint
as well as the client calls for removing a letter
that has not been confirmed as part of a word from the board
and move it back to the players hand.

There is a bug in there at the moment.
When the tile is lifted from the board,
the remove request needs to be sent so that the board
can be redrawn correctly.
However, during redraw, the lifted tile vanishes.