Add scoreboard getter and gameOverHandling
Add the logic to indicate when a game is over.
GameOver is currently defined at the point
as soon as the one of the players has no more letters
in their hand and gannot retrieve anymore letters
from the letter backlog as it is empty.

Add a scoreboard getter to retrieve the names and points
of all players.

As a ui prompt as soon as game is over and display the winner.
Untested.

Todo:
- Testing
- Nicer UI
- Sounds