Add sound efects, fix last row bug
Fix the bug that prevented a letter from the last row
to get read by the "vertical word" algorithm.
The issue was that the wrong index for the bottom border letter int
the vertical word slice was defined.
Apparently a sub-slice is defined so that the right-outer
border index is not part of the sub-slice i.e. a[1:3] includes
indexes 1 and 2 but not 3.

Add sounds effects for letter placements, new game, end game and in/valid word
confirmation.

Todo:
- Wildcard prompt.
- Testing of game over.
- Visual effects for word confirmation (how many points has the player gained).
- Better error messages
- Gerneral UI improvements.
- Background music.