Remove adjacent plaecment checking
The placement legality algorithm was unnecessarily
complicated and too restrictive.

Players are now allowed to initially place tiles wherever they want
(unless the board in completely empty in which case it needs to be the middle tile).

After that, letters can be placed anywhere on the same row or column
as the first two placed letters in one turn.

Needs testing.