Show points gained, ui imrovements
Let the gole server return
the amount of points gained by the player
in a turn and show an animation on the
client side if the turn was successful.

Improve the user interface style.
Show legal placements more subtle
and difference locked and movable tiles
on the board.

TODO:
Wildcard Tiles, client dialog.