Fix issues with gameOver and TW, DW effect
Fix initialization of the ScoreBoard map.
Needs to be initialized before values are
added. No need for a specific data type
instade just make(map[...]...).

For TripleWord and DoubleWord
effect we need to add 1 for double word
and 2 for triple word to the multiplier,
since our base (because it's a multiplication)
is already 1.