Add go get and add index legality function
Add a go get call to the Makefile prepare target
that is called by the pipeline and potentially
by developers to prepare the workspace for building/testing
the souce code.

Add a function to board.go for checking
whether a given set of coordinates is existent
withing the board (i.e. is within bounds).