Copy source to GOPATH
Simply sym-linking the gole
source code to the $gopath/src
directory is error prone
and leads to complications with go commands.

Instead we're going to copy the gole go source to the
gopath source folder and reference it with $gole_path.

This way we can build the binary into the $gopath/bin
directory and call it from there for service start.