client: GUI changes, api: add /replace
Make changes to gui.
The player container as well as the new game button
are moved to the bottom of the window.

The letter struct has now an ID attribute,
so that every letter tile is unique.

There is now a /replace api point
that has already been fully implemented on the
client and server-side.
This endpoint will be called when placing a wildcard
letter on the board and after the player has
chosen a letter with which to replace the wildcard
character.