GUI improvements, wildcard fixes, rearrange hand
Make the electron gui a bit more colourful
and more intuitive to use.
Still a bit work left here.

Use a fontawesome symbol for the wildcard letter
which also required some logic changes on the
client side.

Implement a rearrange player hand
algorithm to sort the letters of the player
in their hand.
Add an API on top on ther server side and implement
the according js client library.

Add the logic to the client.
Sorting will now be called whenever the player
places a letter into the hand (e.g. remove from board or rearrange in hand).
And a shuffle button has been added.