Refactor and fix FinishTurn algorithm
The FinishTurn algorithm is currently
too bloated and needs to be exploded in
more than one function.

There is also work to do from a performance
perspective.
The procedure needs to be optimized
to what is actually required by the game
logi.
This may require teh removal of certain
unnecessary abstractions.