Tidy up FinishTurn procedure
Extract the "findNewWords" procedure
to its own function.
See last commit for closer discrion of the reasons
behind this.

The FinishTurn function is now considerably
smaller and the separated function is less expensive
than the procedure as it was existent before.

Still to do:
It should be possible for players
to see haw many poins they are about to gain
for the turn they are playing at the moment.
This information should be shown
every time the player places a letter on the board.
The validity of the words on the board does not
play a role at this time.