Return pot. points f. words at letter placement
Make use of the GetPointsForWord funtion
in the PlaceLetter function to determine
the potential words a player would gain for
a new word if s/he were to end the current turn.

This point check does not consider the validity of
the new word i.e. no dictionary check is commited.