Add potential points client logic
Let the client collect the potential points
for words after letter placement.
Add algorithm for adding these points to the
client side board tile structs and add a
logic for displaying.

BUG on SERVERSIDE:
If a new tile extends an already locked word,
the letter index of the letter through which
the word has been detected on the server
is not the fist index of the complete word
as these parts have already been locked.
Need to use the index returned by the GetHorizontal/VerticalWord
algorithm insted.