Fix getNewWords algorithm and client-side present.
Fix the issue with the indexing of the words
found in the getNewWords algorithm as described in the last commit.

Also add some more presenting logic.

Todo:
The client-side gameplay is becoming slower hand buggier.
The rendering of the board needs to be changed accordingly
so that loading is faster and so that actions by the user
have the expected effect (e.g. placing of a tile should happen
only as intended by the user).

also:
The potentialPoints should have their own
API endpoint on the server.
At the moment it is not possible to get
the potential points for a word after
a letter has been removed by the player.